Temporary folder renames to fix Vercel build bug.
