+++ components/Attribute-foo/index.tsx
