+++ components/Footer-foo/index.tsx
