+++ components/Pagination-foo/Pagination.module.css
