+++ components/Pagination-foo/index.tsx
